requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.61.1
cryptography>=42.0.0  # Required for django-allauth
Pillow>=10.2.0  # For image handling in UserProfile model
requests>=2.31.0  # For testing JWT endpoints 